+++ htdocs/Office365LicensesExport.php
@@ -11,8 +11,8 @@ use CNCLTD\LoggerCLI;
 use PhpOffice\PhpSpreadsheet\Spreadsheet;
 use PhpOffice\PhpSpreadsheet\Style\Fill;
 use PhpOffice\PhpSpreadsheet\Writer\Xlsx;
-
 require_once("config.inc.php");
+global $cfg;
 require_once($cfg["path_dbe"] . "/DBEPortalCustomerDocument.php");
 require_once($cfg["path_dbe"] . "/DBEOSSupportDates.php");
 require_once($cfg["path_dbe"] . "/DBEHeader.inc.php");
@@ -329,6 +329,9 @@ function processMailboxes(Spreadsheet $spreadSheet,
                 break;
             case 'RoomMailbox':
                 $mailboxes[$key]['RecipientTypeDetails'] = "Room";
+                break;
+            case 'EquipmentMailbox':
+                $mailboxes[$key]['RecipientTypeDetails'] = "Equipment";
                 break;
         }